refactor: move background logic to workers folder
